requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0

lightrag-hku
tiktoken

neo4j>=5.14

pytest>=8.0.0
pytest-asyncio>=0.23.0
black